backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
Flask-CORS==4.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mysql-connector-python==8.2.0
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
psutil==7.1.3
